backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-openai
langchain-community
tavily-python
python-dotenv